Move factory around
