+++ x-pack/plugin/searchable-snapshots/src/main/java/org/elasticsearch/xpack/searchablesnapshots/SearchableSnapshots.java
@@ -5,10 +5,6 @@
  */
 package org.elasticsearch.xpack.searchablesnapshots;
 
-import org.apache.lucene.index.IndexWriter;
-import org.apache.lucene.index.IndexWriterConfig;
-import org.apache.lucene.index.NoMergePolicy;
-import org.apache.lucene.store.Directory;
 import org.apache.lucene.util.SetOnce;
 import org.elasticsearch.action.ActionRequest;
 import org.elasticsearch.action.ActionResponse;
@@ -16,9 +12,7 @@ import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
 import org.elasticsearch.cluster.node.DiscoveryNodes;
 import org.elasticsearch.cluster.service.ClusterService;
-import org.elasticsearch.common.blobstore.BlobContainer;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
-import org.elasticsearch.common.lucene.Lucene;
 import org.elasticsearch.common.settings.ClusterSettings;
 import org.elasticsearch.common.settings.IndexScopedSettings;
 import org.elasticsearch.common.settings.Setting;
@@ -30,26 +24,18 @@ import org.elasticsearch.env.NodeEnvironment;
 import org.elasticsearch.index.IndexSettings;
 import org.elasticsearch.index.engine.EngineFactory;
 import org.elasticsearch.index.engine.ReadOnlyEngine;
-import org.elasticsearch.index.seqno.SequenceNumbers;
-import org.elasticsearch.index.shard.ShardPath;
-import org.elasticsearch.index.snapshots.blobstore.BlobStoreIndexShardSnapshot;
 import org.elasticsearch.index.store.SearchableSnapshotDirectory;
-import org.elasticsearch.index.translog.Translog;
 import org.elasticsearch.index.translog.TranslogStats;
 import org.elasticsearch.plugins.ActionPlugin;
 import org.elasticsearch.plugins.EnginePlugin;
 import org.elasticsearch.plugins.IndexStorePlugin;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.plugins.RepositoryPlugin;
-import org.elasticsearch.repositories.IndexId;
 import org.elasticsearch.repositories.RepositoriesModule;
 import org.elasticsearch.repositories.RepositoriesService;
-import org.elasticsearch.repositories.Repository;
-import org.elasticsearch.repositories.blobstore.BlobStoreRepository;
 import org.elasticsearch.rest.RestController;
 import org.elasticsearch.rest.RestHandler;
 import org.elasticsearch.script.ScriptService;
-import org.elasticsearch.snapshots.SnapshotId;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.watcher.ResourceWatcherService;
 import org.elasticsearch.xpack.searchablesnapshots.action.ClearSearchableSnapshotsCacheAction;
@@ -58,21 +44,16 @@ import org.elasticsearch.xpack.searchablesnapshots.action.SearchableSnapshotsSta
 import org.elasticsearch.xpack.searchablesnapshots.action.TransportClearSearchableSnapshotsCacheAction;
 import org.elasticsearch.xpack.searchablesnapshots.action.TransportMountSearchableSnapshotAction;
 import org.elasticsearch.xpack.searchablesnapshots.action.TransportSearchableSnapshotsStatsAction;
-import org.elasticsearch.xpack.searchablesnapshots.cache.CacheDirectory;
 import org.elasticsearch.xpack.searchablesnapshots.cache.CacheService;
 import org.elasticsearch.xpack.searchablesnapshots.rest.RestClearSearchableSnapshotsCacheAction;
 import org.elasticsearch.xpack.searchablesnapshots.rest.RestMountSearchableSnapshotAction;
 import org.elasticsearch.xpack.searchablesnapshots.rest.RestSearchableSnapshotsStatsAction;
 
-import java.io.IOException;
-import java.nio.file.Path;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
-import java.util.function.LongSupplier;
 import java.util.function.Supplier;
 
 import static org.elasticsearch.index.IndexModule.INDEX_STORE_TYPE_SETTING;
@@ -134,8 +115,13 @@ public class SearchableSnapshots extends Plugin implements IndexStorePlugin, Rep
 
     @Override
     public Map<String, DirectoryFactory> getDirectoryFactories() {
-        return Map.of(SearchableSnapshotRepository.SNAPSHOT_DIRECTORY_FACTORY_KEY,
-            new SearchableSnapshotsDirectoryFactory(repositoriesService::get, cacheService::get, System::nanoTime));
+        return Map.of(SearchableSnapshotRepository.SNAPSHOT_DIRECTORY_FACTORY_KEY, (indexSettings, shardPath) -> {
+            final RepositoriesService repositories = repositoriesService.get();
+            assert repositories != null;
+            final CacheService cache = cacheService.get();
+            assert cache != null;
+            return SearchableSnapshotDirectory.create(repositories, cache, indexSettings, shardPath, System::nanoTime);
+        });
     }
 
     @Override
@@ -165,69 +151,6 @@ public class SearchableSnapshots extends Plugin implements IndexStorePlugin, Rep
             new RestClearSearchableSnapshotsCacheAction(),
             new RestMountSearchableSnapshotAction()
         );
-    }
-
-    private static class SearchableSnapshotsDirectoryFactory implements DirectoryFactory {
-        private final Supplier<RepositoriesService> repositoriesService;
-        private final Supplier<CacheService> cacheService;
-        private final LongSupplier currentTimeNanosSupplier;
-
-        SearchableSnapshotsDirectoryFactory(Supplier<RepositoriesService> repositoriesService,
-                                            Supplier<CacheService> cacheService,
-                                            LongSupplier currentTimeNanosSupplier) {
-            this.repositoriesService = repositoriesService;
-            this.cacheService = cacheService;
-            this.currentTimeNanosSupplier = currentTimeNanosSupplier;
-        }
-
-        @Override
-        public Directory newDirectory(IndexSettings indexSettings, ShardPath shardPath) throws IOException {
-            final RepositoriesService repositories = repositoriesService.get();
-            assert repositories != null;
-            final CacheService cache = cacheService.get();
-            assert cache != null;
-
-            final Repository repository = repositories.repository(
-                SearchableSnapshotRepository.SNAPSHOT_REPOSITORY_SETTING.get(indexSettings.getSettings()));
-            if (repository instanceof BlobStoreRepository == false) {
-                throw new IllegalArgumentException("Repository [" + repository + "] is not searchable");
-            }
-            final BlobStoreRepository blobStoreRepository = (BlobStoreRepository) repository;
-
-            IndexId indexId = new IndexId(indexSettings.getIndex().getName(), SNAPSHOT_INDEX_ID_SETTING.get(indexSettings.getSettings()));
-            BlobContainer blobContainer = blobStoreRepository.shardContainer(indexId, shardPath.getShardId().id());
-
-            SnapshotId snapshotId = new SnapshotId(SNAPSHOT_SNAPSHOT_NAME_SETTING.get(indexSettings.getSettings()),
-                SNAPSHOT_SNAPSHOT_ID_SETTING.get(indexSettings.getSettings()));
-            BlobStoreIndexShardSnapshot snapshot = blobStoreRepository.loadShardSnapshot(blobContainer, snapshotId);
-
-            Directory directory = new SearchableSnapshotDirectory(snapshot, blobContainer);
-            if (SNAPSHOT_CACHE_ENABLED_SETTING.get(indexSettings.getSettings())) {
-                final Path cacheDir = shardPath.getDataPath().resolve("snapshots").resolve(snapshotId.getUUID());
-                directory = new CacheDirectory(directory, cache, cacheDir, snapshotId, indexId, shardPath.getShardId(),
-                    currentTimeNanosSupplier);
-            }
-            directory = new InMemoryNoOpCommitDirectory(directory);
-
-            final IndexWriterConfig indexWriterConfig = new IndexWriterConfig(null)
-                .setSoftDeletesField(Lucene.SOFT_DELETES_FIELD)
-                .setMergePolicy(NoMergePolicy.INSTANCE);
-
-            try (IndexWriter indexWriter = new IndexWriter(directory, indexWriterConfig)) {
-                final Map<String, String> userData = new HashMap<>();
-                indexWriter.getLiveCommitData().forEach(e -> userData.put(e.getKey(), e.getValue()));
-
-                final String translogUUID = Translog.createEmptyTranslog(shardPath.resolveTranslog(),
-                    Long.parseLong(userData.get(SequenceNumbers.LOCAL_CHECKPOINT_KEY)),
-                    shardPath.getShardId(), 0L);
-
-                userData.put(Translog.TRANSLOG_UUID_KEY, translogUUID);
-                indexWriter.setLiveCommitData(userData.entrySet());
-                indexWriter.commit();
-            }
-
-            return directory;
-        }
     }
 
 }
